% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_density_plot.R
\name{generate_density_plot}
\alias{generate_density_plot}
\title{Generate density plot}
\usage{
generate_density_plot(data, outcome_var, outcome_var_label)
}
\arguments{
\item{data}{Data frame that has outcome + column `significant`.}

\item{outcome_var}{String. Name of outcome variable.}

\item{outcome_var_label}{String. Label for x-axis.}
}
\value{
A ggplot object with density plot.
}
\description{
Makes density plot of outcome variable, showing all values + significant ones.
}
\examples{
# prep <- prep_data("hurricane")
# df <- prep[[1]]
# generate_density_plot(df, "edif", "Excess fatalities")
}
