% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_heatmap_strips.R
\name{generate_heatmap_strips}
\alias{generate_heatmap_strips}
\title{Generate heatmap strips}
\usage{
generate_heatmap_strips(data, outcome_var, strip_vars, variable_labels)
}
\arguments{
\item{data}{Data frame with outcome + decision variables.}

\item{outcome_var}{String. Name of the outcome variable.}

\item{strip_vars}{Character vector with names of decision variables.}

\item{variable_labels}{Named vector/list with display labels for decision variables.}
}
\value{
A ggplot/patchwork object with the heatmap strips.
}
\description{
Builds heatmap strips for each decision variable across outcome bins.
}
\examples{
# prep <- prep_data("hurricane")
# df <- prep[[1]]
# strip_vars <- prep[[4]]
# variable_labels <- prep[[5]]
# hm <- generate_heatmap_strips(df, "edif", strip_vars, variable_labels)
}
