% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiverse.R
\name{plot_multiverse}
\alias{plot_multiverse}
\title{Plot multiverse}
\usage{
plot_multiverse(use_case = "hurricane", output_path = NULL)
}
\arguments{
\item{use_case}{String. Use case name.}

\item{output_path}{Optional string. File path to save combined plot.}
}
\value{
A ggplot/patchwork object with combined multiverse plot.
}
\description{
Prepares data for a given use case, builds density plot and heatmap strips + stacks both.
}
\examples{
# p <- plot_multiverse("hurricane")
# p
}
