% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare multiverse data}
\usage{
prep_data(use_case)
}
\arguments{
\item{use_case}{String. Name of the use case.}
}
\value{
A list: (df, outcome_var, outcome_var_label, strip_vars, variable_labels)
}
\description{
Loads and prepares data for a given use case.
}
\examples{
# prep <- prep_data("hurricane")
}
